+++ donkeycar/utils.py
@@ -261,7 +261,7 @@ def gather_tubs(cfg, tub_names):
     takes as input the configuration, and the comma seperated list of tub paths
     returns a list of Tub objects initialized to each path
     '''
-    from donkeycar.parts.tub import Tub
+    from donkeycar.parts.datastore import Tub
 
     tub_paths = gather_tub_paths(cfg, tub_names)
     tubs = [Tub(p) for p in tub_paths]
